Add express-template-hot to starter template list
Express template hot is a copy of express template which separates nuxt and the api server in development mode, but connects the two using a proxy attached to the nuxt server. This prevents file changes from breaking hot reloading of nuxt and the api server reloads faster via live reload as well because it does not need to compile nuxt. You may read more in the backpack readme about the issue.